+++ src/common/providers/configProvider.ts
@@ -118,6 +118,7 @@ const decorationOptions = {
     delayTracking: undefined as boolean | undefined,
     soundPlayback: undefined as boolean | undefined,
     specificSoundPlayback: undefined as boolean | undefined,
+    createMetaskillDocumentation: undefined as boolean | undefined,
 };
 
 const editorConfigCache = {